feat(rows): turn a row on or off from its own editing page
The switch lived only on the Rows card, and the editor's own copy sent you back there for it —
a page change to do the one reversible thing to the row you already have open. It now sits beside
Run now, sharing one component with the card so the two cannot drift apart, and keeping the
confirmation that turning a row off takes it off Plex for everyone at the next run.

Removing and deleting stay fenced off at the bottom: they reach into other people's Plex and Cancel
does not undo them, so they have no business one pixel from "rebuild this row". But fenced off had
become invisible — nothing on the first screenful said they existed — so the header now carries a
quiet "Remove or delete…" that scrolls to them.
